cleanup, refactor paths taken display
